Compiler warnings on Windows
git-svn-id: svn+ssh://svn.code.sf.net/p/libjpeg-turbo/code/trunk@914 632fc199-4ca6-4c93-a231-07263d6284db

+++ djpeg.c
@@ -573,7 +573,7 @@ main (int argc, char **argv)
         else
           fprintf(stderr, "%s: can't read from stdin\n", progname);
       }
-      insize += nbytes;
+      insize += (unsigned long)nbytes;
     } while (nbytes == INPUT_BUF_SIZE);
     fprintf(stderr, "Compressed size:  %lu bytes\n", insize);
     jpeg_mem_src(&cinfo, inbuffer, insize);
